Split up test_api.py
