+++ backend/tests/unit/test_moves_bishop.py
